Splits utils into multiple submodules.
